refactor : confirmPaymentService 중복 로직 개선
